Add metadata to pack files
Pack files now contain a section at the end of the file that contains
the metadata describing the blob content.

The last 4 bytes of the filecontain the length of the section. For each blob
in the pack, the metadata section contains:

  - ID (32 bytes)
  - Length (4 bytes)
  - Blob type (1 byte)

The structure of the metadata section is:

  | id | len | type | id | len | type | ... | META SIZE |

This metadata should allow to scan the contents of the pack directly, without
reading the index. Also, if a pack became unreferenced from any snapshot, its
contents can still be processed.

+++ src/repository/gc.rs
@@ -82,8 +82,8 @@ pub fn scan(repo: Arc<dyn RepositoryBackend>, tolerance: f32) -> Result<Plan> {
         if !plan.referenced_blobs.contains(id) {
             pack_garbage
                 .entry(locator.pack_id)
-                .and_modify(|size| *size += locator.length)
-                .or_insert(locator.length);
+                .and_modify(|size| *size += locator.length as u64)
+                .or_insert(locator.length as u64);
             spinner.inc(1);
         }
     }
@@ -188,8 +188,8 @@ impl Plan {
                     let data = self.repo.read_from_file(
                         global::FileType::Object,
                         &pack_id,
-                        offset,
-                        length,
+                        offset as u64,
+                        length as u64,
                     )?;
                     self.repo.save_blob(
                         blob_type,
